Tabular simulation information
Laid out basic simulation info as a table in the simulation overview.
I will not continue this work, before I wrote a class to model the
simulation data. This way whereever we get the simulation data from -
form, json as argument, or from the database - it'll have the same
interface. And that will make it so much easier to display.