Started translating the db-structure into SQL-Alchemy ORM
With this commit I made efforts to translate the plain SQL code into
SQL-Alchemy ORM code.

The biggest change here is, that we no longer have each table be created
once for each model, but now having only one set of tables for all
models. To distinguish which model an entry belongs to I added a `model`
column to each table that saves just that. This also eleminates the need
for using schemas to organise tables.