Split `db/db.py` into one file for each table
With this commit I split up the `db/db.py` into one file per table. I think
smaller files are easier to work with and managing imports is easier as
well.
In addition there is the `db/base.py` where the database connection gets
initialized and the base for all tables gets initialized.
The `Models` enum containing all valid model names was moved out of the
`db` package, as it does not directly belongs to the db. It was also renamed to `SimulationModel` and now lives in a new file `values.py`.
In the `db/__init__.py` we also now create all tables in the database.