Safer SQL query in SimulationManager
In the SimulationManager, we query the database to obtain the newest id
so we can then load the image for that very simulation. But as we setup
the database in the `SDTM.py` the database might not be initialized when
we query for the id in the SimulationManager.
And as we not yet have a working ORM setup, I just added some try-except
statements so that any error thrown by the database doesn't bubble up.
This is not the way to go, we just need a quick fix.