Renaming tables and columns
for even more consistency I renamed the occurance of the field `model`
into `type` to have a clear naming throughout the project. Models are
now what stores the JSON modelling the processes we want to simulate.
Their `types` determine if they are 'hog', 'ion', 'volume' or 'combined'.
Furthermore I changed the singular naming of `Parameter`, `InitialValue` and `OrresuEquation` to being plural. I think this makes more sense as they don't contain one single value, but a set of values.
And last but not least I reworked the `Model` tables column names and
added some colums that I plan to expose on the interface later.
Same with the `Ex` table. It also got some extra fields that will be
useful for statistics later on.