Parameters, InitialValues, OrresuEquations versioning
I realized that when rebuilding the database structure in using the ORM
wrapper from SQL Alchemy, I made a mistake assuming that the column `seq`
on the tables `Model`, `InitialValues` and `Parameters` served as an id. But in
reality it was a kind of versioning that helped to distinguish the
different sets of values from each other. So I added a `version` column
again to reproduce this behaviour.
Doing so I no longer can have the forein key relation from the `Ex`, so
I removed it for now. We should come up with a way to model this
relationship again at a later poinDoing so I no longer can have the
forein key relation from the `Ex`, so I removed it for now. We should
come up with a way to model this relationship again at a later point.