Fixed simulation failing when combined models are selected
As I am trying to use an enum wherever we have to set the type of model
we want to run and some parts of the code still use a plain string, we
ran into errors as the naming between those two is slightly different.
With passing the string value into our enum, we make sure to get the
right value and things work again.