Finished extracting DataExtraction from `SDTM.py`
With this commit I finished the extraction of the `DateExtraction`
class. Things are no longer broken, meaning I can run simulations
without getting errors and images are being displayed as well.
This work helped me much with understanding what is actually needed for
a simulation to run. Also it lays the gourndwork for modeling the simulation
process with classes rather then just a blob of code.