Fixed images not showing up
With the last commit I broke the image results showing up in the
web interface. This commit fixes that. It does so by addressing the
issue of querying the next id to beinserted into the `Ex` table. This is
not a good practise as you will run into issues when running multiple
simulations concurrently.
By giving the Simulation instance in the SimulationManager her uuid back,
which we then pass to the process in which the actual calculations run
in, we can have the simulation process add the uuid to the database. If
we then want to access the simulation data again from within the
SimulationManager, we can query it using the uuid.