[Broken] Importable sdtm
With this commit, the `SDTM.py` can now be imported and the simulation
executed via calling the `sdtm` method.
This way we no longer need to execute a bash command to start up the
`SDTM.py` and pass the arguments as a JSON string.
However doing so, we run into troubles with the database. Ocasionally it
crashes and throws an error. Also, sometimes simulations get stuck and
don't stop running - at least for as long as I had the nerve to wait.
These issues will be adressed in the following commits.
Simulation results are however unaffected.