Initial values and parameters now enclosed by sets
Before the initial values and parameters were just loosely flying around
in the database, making it hard to constrain their relation to an
exposure, leaving potential for errors.
Now all initial values and parameters are enclosed by a set
(`InitialValueSet` and `ParameterSet`). The sets are easier to reference
and so we can constrain the relation to the exposures. The values for
the sets are stored in an extra table, similar to the way they were
stored before. But now we also can constrain them to their respective
set.
This makes working with the database easier and less error prone. Also I
like the new naming. It is much clearer and we don't need to be
uncertain about whether its plural or singular. Now it clearly is
singular, how it should be.