Extracted initialisation procedure into own model
In this commit I moved all the code that prepares something so the app
can run into one file `web_interface/init.py`. There we initialize the
database and create folders that might not exist.
This is the preparation for installing logging in the project. Also all
those preparational code now has a place to live.