"Role root does not exist" fixed
In this commit I fixed the repeatedly occuring error message on the
postgres image, that the role root would not exist.
The fix was actually quite simple. The error was not with postgres, but
with the docker healthcheck that polled the database periodically. I now
added the correct username to the healthcheck and the message no longer
appears.