'Role root does not exist' fix
In this commit I try again to solve the error appearing in the log of
the postgres container. Testing showed that the healthcheck defined in
the `docker-compose.yml` is to blame again. I patched the healthcheck so
that it uses the correct user now. Also I tweaked the other healthcheck
values to hopefully give the database more time to initialize and not
make the healthcheck fail so fast. I hast gotten a bit stricter though.