Complete database model for simulation
We have implemented our ORM wrapper but we are still not using it to its
full potential. So the goal in this commit is, to remove some of the
many classes flying around in the project and unify them in our ORM
wraper classes / models. This way I hope to some day mostly work with
ORM wrapper instances in the code.

I extended the database model so it can store all information that were
inputted into the web interface. Not only gives us that a more complete
image of the simulation. We also can now rerun previous simulations.
Specifically, I added the hog specific attributes 'signal-type' and
'NaCl-impulse' and the specified Stimuli and Impulses.

I completely overhauled the SimulationManager, removing most of the
classes defined there, as they are now represented in the ORM models. I
moved some of the code creating the arguments for the simulation (`SDTM.py`) from there into
the Ex-model and reworked the way the form input was transformed into
classes. The form input is now directly fed into the ORM model.

`SDTM.py` now gets the ORM model for a simulation passed into it. This
in used throughout the code to add values from the simulation to
complete its data set.

In the database package I removed the close statement for a session in
the sessionScope. I needed to so this as I would otherwise get an error
in the template files, that my ORM instance is no longer bound to a
session, thus I couldn't access all of its data. This is rather a quick
fix. We need a better way of handling sessions in general.

I also adapted the templates to work with the ORM model of the
simulation.