Prevent double initialization on startup
In debug mode, where the code is realoded on change, we have the problem
of flask causing our project to initialize twice at the same time.
Usually, resulting in errors from the database.

Why Is that?
In debug mode, flask reloads the code on change for convenient
development. For that to work, there is an ever running watch-process,
watching the code for changes. On every change, the watch-process
starts a new server-process that then actually funtions as the server.
On startup however, the watch-process runs the code, triggering
the database initialization. At the same time, it starts a server-process
to work as the - well - server, which also runs the code and initializes the
database. And thus, you usually get an error when starting up the project with
an empty database.

In this commit I introduced a check, that prevents the code from
running in the watch-process.