[Broken] Working task queue prototype
This is just a prototype, so no actual simulation will run.

In this commit I build a working prototype for scheduling simulations and
asynchronusly saving their results. This was quite a huge change to
implement.

I needed to add the simplejson package as I had issues with the default
one, not being able to serialise the simulation data. Thus the
Dockerfile update.

I added three more services. RabbitMQ for processing the messages; One
for running the simulation in and another one for storing the results in
the database. I probably could have integrated the last one into the web
service. But as I already experienced so much issues with SQL Alchemy
and threading / subprocessing, I decided against that.