[Broken] Removed any database interaction from `SDTM.py`
In this commit I made efforts to remove any database dependent code from
the `SDTM.py`. This way the simulation will be more decoupled from the
rest of the code and can be replaced by the newer version written in
Julia.

I removed every database query and replaced it with the data provided by
the task queue. All writing queries were transformed into populating a
dictionary which is returned at the end of the simulation. The dict will
then be passed on to the `result-worker`.

Furthermore, the simulation no longer loads the model json from the
disk, but uses the one supplied by the task queue.

I also cleaned up the imports, especially the evaluated ones in
`SDTM.py`.