[Broken] Only pass necessary information through the task queue
In this commit I reworked the simulation worker, to still function, when
only receiving the minimal data it needs. Before, we sent all impulses
and stimuli - even the ones that were unrelevant for the simulated
model. Now, we pass only the impulses and stimuli needed for the
selected model to do its work.
This required some changes to the simulation code. I used the chance to
also cleanup the code a bit.

I removed the calls to the `locals` function. They are no longer needed
and have no apparent effect.

I also needed to tweak the code a bit to make hog run with the reduced
set of information.

I moved some funtions over into `SDTM.py` to clean up the code where we
listen for the simulations passed by the task queue. This made it
possible to only pass the raw data received from the task queue directly
to `SDTM.py`.