Using float rather than double precision
I removed all use of double precision from the code and replaced it with
floats. This way JSON serialization works without special packages and
it allowed us to remove various castings to float from the code that
were necessary to make the simulation work. Also I think this will
improve compatability with other simulation workers in the future.