Moved database initialization into the db package
In this commit I moved all the data needed for initializing the database
into the db package. That is the json files describing a model, the
initial values and parameters.
Doing so I created a few new classes to model the initial data. This
makes working with the initial data easier.

To insert the data, I use SQL Alchemy events. This way I made
sure, to only insert initial values, when SLQ Alchemy just executed
create statements for initializing the database tables.