Db initialization completely moved
With this commit all old data for initializing the database is removed.
THe initialization now is completely handled by the `initialData`
package, a sub package of the `db` package.

For all initial data to work 100% I needed to adapt the initial data
model `ModelData` to work not only with constant values, but with
expressions as well. For that I added the option to pass equations,
describing initial values or parameters as a string. These strings will
be not be calculated, but persisted in the database just as they are.
To ensure that these equations can be calculated, I added a precedence
column to the `InitialValue` and `Parameter` model to ensure they are
evaluated in the order they were defined. I defined an ordering for the
relationships between the `InitialValueSet`, `ParameterSet` and `InitialValue`,
`Parameter` model. This way the values will be always retrieved in the
correct order to ensure that equations can be evaluated.

The `Solver` in the `simulationWorker` package was adapted to work with
the values from the database, rather than loading these values from
disk.
As initial values and parameters are now stored as strings in the
database I needed to specifically tell the panda data frame to use flaot
as data type and not infer it.

Lastly, I removed the code and related dictionary from the simulation
that exports the terms to the database. The option was disabled
hardcoded, so I deemed it to be no longer relevant.