Separated Code from Configuration
With this commit all code files were moved into their own folder, named
`projectQ`. This way they are separeted from the configuration files in
the root of the project.
The forlder was named `projectQ` to make it importable using python
under its real name, rather than using `src` or something.

To make things work again, I adapted the mounting of the code in the
`docker-compose.yml`.

I also adapted the `.gitignore` to make sure the generated pictures will
be excluded from the repository.