Removed temporary docker image
When implementing the first prototype of using a message broker for
communication between services, I had some troubles with JSONifying the
messages to be send over the broker. They had decomals in them which
could not be serialized. So as a quick fix, I installed the simplejson
package. But as I didn't want to use it and rather move to using floats,
I just created a temporary docker image tag.
unfortunately I forgot to remove that tag from the `docker-compose.yml`
when I was done moving to using floats.