Database username and password in `.env`
With this commit I extended the `.env` file to also hold the username
and password of the database.
The healthcheck was adapted to also use the environment, rather than
hardcoded values.
The `db` package was adapted to use the environment values and no longer
hardcoded ones.
The environment values are retrieved in the `values` module.

With these values living in the `.env` file we are getting a bit closer
to actually being able to deploy the system ona a server somewhere.