Advanced networking
I created two networks, one for talking to the database and one for
talking to the message broker (taskQueue). This way the services are
more separated and can only access the resources they really need.
This is not really needed for the servoces to function, but rather to
prevent strange interactions between the services.