Refactored `docker-compose.yml`
In this commit I took the time to make the `docker-compose.yml` pretty
again. I rewrote all comments in english and tried to explain every
attribute used in the file. Also, I ordered all attributes by name, to
make navigating the file easier. Furthermore, I renamed the taskQueue
service to 'messageBroker' as this is what it does. Renamed the networks
to reflect the name of the service they are created for.

Also I some more values to the `values` module containing the host name
of the message broker and the host name of the database.