Version pinning and updated Dockerfile
In this commit I pinned the versions of all packages and images we are
using in our project.
Also I updated the docker file to use debian instead of alpine linux.
With debian installing packages is much faster as they can be downloaded
pre compiled. I mainly did this to speed up the image building as I
otherwise couldn't efficiently test the changes I made to the
Dockerfile.