First step towards react based web app
I moved the flask code from the `web` folder into a `server` folder as
this is what flask will be used as from now on. The `web` folder will
contain all the code that makes up the react based web-app.
Also I installed flask-socketio in the dockerfile as preparation for
what is coming.