Set up configuration for the web app
In this commit I created all configuration files needed for the web app
to run properly. This includes, the `package.json`, `package-lock.json`,
`webpack.config.js` and `.babelrc`.
I adapted the `.gitignore` to exclude certain Nodejs / React specific
files and folders.
I updated the readme with the instructions on how to run the project now
and how to deveop the web app.
I created a new template that is used to serve the web app and adapted
the current web interface to link to the web app. The plan is to have
both running besides each other until the web app has all the
functionality of the current web interface.
I moved the generated pictures into the server package. There now is a
`static` folder for static assets like thise pictures. The bundled code
lives there as well.