First working connection with SocketIO
In this commit I established the first working websocket connection between the server and the web app.
For this I had to make changes to the server so that we no longer start
Flask, but Flas-SocketIO. This turned out to be quite tough. I needed to
make some significant changes to the server package. Most notably the
server is no longer started using the Flask CLI. This did not work with
debugging enabled, more specifically the code reloading. Now the server
is started as a package unsing a plain python command. This works, only
the logging is somewhat broken. Connection logs do not appear until code
changes reload the server.
I updated the `Dockerfile` installing `eventlet` for true websocket
connections.
In the web app I installed the `socket.io-client` and just adding some
code to connect it to the server.
So far no messages are being exchanged, only the connection is
established.