Implemented simulation overview
In this commit I implemented the simulation overview in the
web app. It looks quite similar to the one in the old web interface, I
just updated some minor things.

The following things I did to make it happen:
I installed Flask-RESTFul to have an REST api that the web app can query
data from. (SocketIO will be used for live updates only.) For that I
updated the Docker image. Its tag now has the branch name in it. I don't
like pushing images with tags that look like they are main releases from
the master branch.
Also, I expanded the `Ex` class to now also return a brief dict
containig a limited set of its data. This is basically all data that is
needed to just display the simulation in a list of simulations. The
methods that retrieve the simulations data as a dict were chagnes to
incoorporate an option to return a json-ready dict. That is a dict that
can be serialized with the standard json package.
In the web app I wrote a ton of code that makes up the new simulation
list. Installed a few packages to make things work. Updated Bootstrap to
have loading animations. Inplemented JSS in some places and started
using webfonts to make the UI look a bit prettier.
One more thing to note is that I extracted the Layout out of the
`App.jsx` files App component. Now every component that renders a whole
pages content needs to wrap itself in one layout to not look broken.