Fixed simulation image result path
Somehow the simulation stopped working as it would no longer save
pictures. This is because I changed the path of the code mounted into
the containers. So the error should have shown up there already. It
didn't though. Or maybe I forgot testing it.

Anyways, now the path is fixed and hopefully more robust as I use the
`ROOT_DIR` constant from the `values` module.