Event System Prototype
In this commit I build a basic event system that relies on our RabbitMQ
message broker.
I implemented an exchange that will handle all events in the project.
Thus it'll be the single point of truth.
I also started working on a decorator based syntax for the messageQueue.
In future I just want to add decorators to functions in order to make
them listen to specific events. So far only one decorator can be used at
a time as they are still blocking.
Also I just wired up the web app to receive the currently only event
that can happen - when a simulation was persisted in the database, after
it was simulated.
The plan is to then update the simulations in the web app accordingly.