overhauled simulation overview
With the new fancy event based communication in our project, we already
register changes in the simulation state in the web app. So why not show
it to the world?

In this commit I overhauled the simulations overview. Instead of two
separate areas, there is now just one big list with all simulations. To
distinguish their status, a small icon was added to the right of their
box indicating whether its waiting for resources, running or finished.