Moved to immutable state for simulations
In this commit I switched from using plain js objects to immutable
OrderedMap and Map for usage in the simulation state. This makes
changing the simuations in the reducer more expressive. Also it allowed
me to re implement sorting of the simulations in their overview.
Due to the event based nature a new simulation can be dynamically added
to the web apps state. To still display them in order they were
created, we need to sort them before displaying. This is much easier
with immutable - as it would be with any library if the alternative is
to implement it yourself.

Tests were adapted.