Improved message queue decorator
In this commit I worked on improving the message queue usage by moving
the async code execution inside the `on` function. This way using the
decorator automatically starts an async listener. Much easier to use
now. Quite like I envisioned it.
Also, messages emitted using the message queue will noe be wrapped in a
default event dict. This dict contains the message as `payload` and some metadata about the
time when the message was emitted.

Also I extended the dbWorker to react on multiple events and keep the
database in sync with what is happening. It now reacts whenever a
simulation is scheduled, started and finished py noting down the time in
the database, making use of the new event structure of message in the
message queue, reading the time directly from the events metadata.

Oh and more consistent logging in both workers.