Scaling wokers works again
with implementing the event based communication in the project we lost
the ability to scale services - meaning to have multiple containers of
the same service to distribute the work amongst them.
In this commit I reintroduced this ability to the event based
communication.

Why did it work before the event based communication?
Before the event based communication system was in place, we just had
two queues, one to hold the scheduled simulations and the other to hold
the reults of a simulation.
Queues are used to distribute messages mongst multiple consumer, that is
clients, containers. So we could just start up as many containers of a
service as we wanted and they all would use the same two queues.

Why did the introduction of the event based system broke the ability to
scale services?
The event based communication system broke the abiblity to scale
services, as it changed the way we declared queues. We started to declare queues
anonymously. This meant the queue had an automatically generated unique
name. Now, whenever we registered a listener for an event, this would
also create a new individual queue just for this one listener. Just
because it had a new, never before seen name. Having multiple containers of
one service would mean that each container would have its own queue. So as
never two containers were connected to the same queue, distributing of
messages would naturally not occur.

How did I reintroduce task distribution in the event based communication
system?
To again be able to distribute messages amongst a pool of containers,
all containers of one service must share the same queue.
To achieve this, I changed the way we created our listeners, adding the
options to pass a service name. When creating a listener with a service
name, we now create a queue that is named like the passed service name
plus the name of the event.
This way having multiple containers of the same service, they will all
register listeners with the same service name and thus share the same
queues. The queue names also contain the event name, so we don't need to
manually separate the messages on the consumer end as each queue just
gets one specific type of message anyway.

This might not be the most elegant way to deal with this problem. I
think having less queues and separating the messages manually in the
consumer could improve performance. But we are not at the point to worry
about such things. First things need to work, performance comes after
that.