Web app now working served from Flask as well as webpack dev server
In this commit I generalized the web app to work when served from flask
as well as when being served from the dev server. This way it does not
matter how you start the project, the web app will work either way.

To achieve this, I firstly declared the host name in the `values.js` to
be used throughout the project. No more hard coded urls! Now the web app
always uses the url it was opened from.
But now, when developing using the webpack dev server, the web app could
no longer connect to the servers websocket as the dev server sitting
in between does not forward websockets. I installed the DefinePlugin to
expose to the web app whether it was being served from the dev server or
not.
Now, when the web app is served from the dev server, it limits itself to
use polling rather than websockets. As it just uses standart http
requests for polling, these request are forwarded by the dev server and
the web app can talk to the flask server.