Simplified startup routine
In this commit I simplified the actions needed to start up the project
and interact with the web interface or the new web app. Basically,
nothing changed to the way you used to start up the project.

For this I added a service using the official `node` image from docker hub.
This service will run once when you execute `docker-compose up` and
bundle the code from the web app so that it can be served by the flask
server.

If you want to developt the web app, an alternative startup routine is
needed, see the readme for details.