Error handling when simulation fails
In this commit I started adding error handling to the project by
introducing a new event that is emitted when a simulation fails.
The status is persisted on the simulation and displayed in the web app
with a red cross. The simulation will be discarded from the message
broker and thus not clog things up.