Redux like event emitting
In this commit I defined the events that can be emitted via the message
broker. I orientated myself on the way actions are defined in redux.
The messageQueue now has `eventTypes` as single point of truth for event
types. Also, there are now `eventCreators` as well. These should be used
to create new events to be emitted.
This allows me to do some basic checking of arguments passed into the
event creator functions and thus we can catch faulty events earlier.

I installed and used `schema` to validate dicts passed into the event
creators.