Class model for message broker events
In this commit I implemented a class model to model the events that are
used to communicate via the message broker. With this in place listening
to messages is more expressive as the message must not be parsed and
checked in the listener. But more importantly this way we have a system
in place to verify outgoing and incoming messages.

The class model for the events lives in the messageQueue package. The
base Event class holds all the functionality, while the subclasses just
define their routing_key and their payload schema. Creating new events
is very easy this way. All events are tested using pythons unittest
package.

For the SimulationScheduled event, I needed to verify that the structure
of the payload matches the structure of our database model `Ex`. For
that I implemented schema definitions on the database models. The
SimulationScheduled event now uses the database schemas to verify if all
necessary data is inside the payload. The `Ex` classes schema
definitions are tested as well.

I adapted the messageQueue `on` and `emit` method to work with the new
Even classes. This was quite the delight.

I adapted all the places where communication with the message broker
took place to use the new events. Much nicer now.

We now have a regex to verify RFC3339 datetime strings in the values
module. This is used in the schema definitions of the events.