Renamed `messageQueue` to `eventSystem`
In this commit I did what I should already have done: renamed the
`messageQueue` package as it was no longer just a package for
interacting with queues, but became a real event system when I
redesigned it.

Now everythings renamed and much nicer organized. I like it.

I also found a bug in the server where I forgot to set the `json_ready`
flag for the messages emitted by the socket.