Fixed the initialization of the result image dir
For the web-app I moved the folder for static files into the server
package. Doing so, the simulation result images were also moved there.
However, I forgot to update the initialization process of that folder.
It was still creating the old folder, rather than the new one.

I changed this in this commit.