Added `Makefile` for easier execution of tasks
In this commit I created a `Makefile` to serve as a collection of tasks
one would like to run. At the moment it is basically jsut a wrapper for
docker commands. But the `clean` target already proved useful. I
discovered a flaw in the initialization code, which does not initialize
all needed directories. I fixed that.
Also I realized, that the bundler service does not install npm packages
and thus would fail when run on a brand new instance of this project.