Added testing make target
In this commit I introduced a make target for testing the web apps code
as well as they python code and running the project clean to see if all
initialization works.

This is my consequence for forgetting to test the updated initialization
in the last few PRs.

So far the tests are run inside docker containers, not really
performant, but I could not get the python ones to work locally. Thanks
matplotlib.