dummy Julia simulation worker; two simulation services
In this commit I started working on the basics required for integrating
the julia simulation into the projects infrastructure. The plan is to
prepare everything, so that we can merge this branch into the
`julia-solver` branch. Here is what has changed:

Firstly, we now have two simulation services `py_simulation` running the
old python simulation and `julia_simulation` for the new simulation
written in julia. By default the project will use the new
`julia_simulation` service as simulation worker. Instructions on how to
use the old, python one can be found in the readme.
The new `julia_simulation` service is running in the official julia
docker image. I first planned on using python for communicating with the
message broker and passing the simulation data to some julia function,
but as this would require a mixed docker image having both julia and
python installed, I didn't further pursue this idea. Though, I don't like the
idea of having to maintain the message broker communication in two
languages either. Perhaps in the future I'll investigate creating a
mixed docker-image.
So far the `julia_simulation` service just connects to the message
broker, just as the old `py_simulation` would. It can receive messages -
scheduled simulations that is - but it neither starts a simulation (how
would it without the code merged), nor is it able to emit simulation
related events, like when it starts the simulation, to let the other
services know about the simulations status.