Abstracted message broker listening
In this commit I recreated the python `eventSystem`s `on` decorator in
julia code for esier and more espressive registration of listeners in
julia. It works almost the same as the python package one.
Declaring anonymous queues is the only thing that is not supported in
contrast to the python packages `on` decorator.