Simulation worker now emits status events
In this commit I added an `emit` function that is baically a copy of the
one found in the python `eventSystem` package. With this, the julia
simulation worker now emits status changes, just like the python one
does.

I also added a dummy implementation of the simulate method as I needed
it to test and implement the `simulation.failed` event.