Event creators
In this commit I moved the creation of the event dictionaries, that will be
emitted to indicate status changes in the process of simulating, into
their own file. There now is a function for each type of event. Calling
these functions with the parameters they require, they will return you a
dictionary that can be emitted via the message broker.
It is quite redux like as I wanted a clean, yet simple solution. I
didn't want to go all out building a fully fledged event system like I
did in python. Especially, when considering that we might use protobuf
in the future for that.

With this the julia simulation worker does not look so bloatet anymore,
but simple and expressive.