Started reqorking the julia simulation
In this commit I started on reworking the julia simulation. I
reorganized the methods used for the simulation and extracted code used
for creating the SDE problem as well as for actually solving the
problem. I ditched everything that I didn't immediately understood. That
is especially the neuronal network matrix. I
hope to add it back in later with the help of @JanPiotraschke.
The simulation now has less code and looks cleaner and finishes! Though
I cannot confirm if the results are correct.

I added a lot of comments which are probably mathematecally
questionable, but hopefully give you an idea of what is going on.

I also added a test to test the dynamic derive function. I hope to
extend the tests to cover all of the simulation. This would help a lot
when making changes to the simulation.