Retrieving all results of the simulation
Thanks to the explanation from @JanPiotraschke I think I understood the
usage of the - I think it is called - `reducer` and the `n` parameter.
With my new knowledge I implemented it straight away. Now rereceive
results for all substances in the simulation.

What I found strange, however, was that I still could supply a wrong
value for `n` - that is not matching the length of the result matrix -
without getting an dimension mismatch.