Custom julia docker image
finally, we also have our own julia image. That is finally we have
pinned versions for all the packages we are using and finally we no
longer need to wait on julia installing and precompiling things,
when started up.
The project now also has the files `Project.toml` and `Manifest.toml`
in which the version for each directly and indirectly needed package are
set. This makes things much more robust as we now have th control over
updating the packages.
With these files in place, we can install and even precompile all
packages when building the docker image. No more hard coded package
installing inside any julia file. It is a real pleasure to just use
`using <Packagename>` at the beginning of a file and that's all. So I
directly removed all usage of `Pkg` from the files that make up the
service.

As we now have two docker images, I separated them by using different
tag names. NOt so sure yet, if I like th idea of having two completely
different images in the same repo, just using different tags.