Tests for volume model
In this commit I extended the simulation tests with an actual model from the
project, the 'volume' model. The data used in the tests is just as it would
come from the web interface.
For comparision I also added a hard-coded version of the volume model to
the tests.

I also added some more `@info` statements for better debugging.