Pre-simulation initial value and parameter evaluation
In this commit I changed the simulation so that it evaluates its given
initial values and parameters to actual numeric values in order to
avaiod errors in used packages.

I also added a test for the `simulate` function where we compare the
results of it, with the results of a hard-coded simulation. For that I
adapted the return value of the `simulate` function for easier
comparision of the results. This will be only intermediary until the
`simulate` function works with all simulation types.