Adapted Volume model test
In this commit I reworked the test for the volume model. It now has its
own file and uses the extended simulation interface where you can
specify a noise level as well as the seed.

With the improved handling of the initial values and parameters
(evaluating them to numeric values before the start of the simulation)
the test worked right away.