Ion simulation test
In this commit I added code to test the Ion simulation.

I needed to adapt the data structure used in the simulation. Everything
is now a list. This way we can ensure that constant values are evaluated
before equations that depend on those.
With the dictionaries which we used for initial values and parameters
before, we could not ensure the correct order of evaluation and thus
running into errors.

The Volume simulation test was adapted to also use the new data
structure.