Moving files
starting the restructuring, I moved almost every file around. Every
programming language now has its own directory, where the code is
separated into packages, applications and services.
The services are reflected
in the `docker-compose.yml`. In the services folders there lives all the
code to make a service run as well as the configuration needed by the
(docker-)service.
An `application` folder is used when the code is service
like, but not reflected in the `docker-compose.yml`. This is the case
for the web-app. It is not a service, as it is served via the flask
server (at the moment).

The julia code-base was separated into the afore mentioned directory
structure. Files were renamed and no longer needed ones were deleted.

Paths were adapted in the `docker-compose.yml` and the `Dockerfiles`.

Also one pycahce file was removed as it mistakenly got commited into the
repo at some point.